Debug time breakdown, modularising charts in time dashboard
